Ajustes : corrigido relacionamento JPA, endpoint boletos/{id}

+++ moduloboleto/src/main/java/br/com/moduloboleto/repositories/BoletoTrackingRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.moduloboleto.entities.BoletoTracking;
 
-public interface BoletoTrackingRepository extends JpaRepository<BoletoTracking, Long>  {
-
+public interface BoletoTrackingRepository extends JpaRepository<BoletoTracking, Integer>  {
 }
